Move listing functionality into container and add routing.
