+++ client/react/containers/app.jsx
@@ -3,6 +3,7 @@ import { Route, Link } from 'react-router-dom';
 import Navbar from '../components/navbar';
 import Home from './home';
 import SignUp from './signup';
+import BrowseListings from './BrowseListings'
 import { history } from '../../redux/store';
 
 const style = {
@@ -18,6 +19,7 @@ const app = ({
         <main style={style}>
             <Route exact path="/" component={Home} />
             <Route exact path="/sign-up" component={SignUp} />
+            <Route exact path="/listings" component={BrowseListings} />
         </main>
     </div>
 );
